+++ lib/l10n/app_localizations_ja.dart
@@ -1077,4 +1077,7 @@ class AppLocalizationsJa extends AppLocalizations {
   String repoCreated(Object repoName) {
     return 'Repository \'$repoName\' created successfully';
   }
+
+  @override
+  String get complete => 'complete';
 }