+++ lib/l10n/app_localizations_pt.dart
@@ -1080,4 +1080,7 @@ class AppLocalizationsPt extends AppLocalizations {
   String repoCreated(Object repoName) {
     return 'Repository \'$repoName\' created successfully';
   }
+
+  @override
+  String get complete => 'complete';
 }